Refactoring: introduce AbstractSemanticIndexManager
