+++ src/langium-model-server/semantic/semantic-types.ts
@@ -3,3 +3,7 @@ export type Valid<T> = T & { __semantic: 'valid' }
 export type SemanticIndex = {
     readonly id: string
 }
+
+export type ModelAwareSemanticIndex<SemI extends SemanticIndex> = SemI & {
+    readonly model: object
+}
